Update organisms.md
Here I just added the name of the process of when the surface tension is reached and I expanded on how the mechanism works because applying a downward rotational force is hard to grasp without a good comparison. So I added a note on how the mechanism is similar to jumping in that a downward force stores energy before being propelled upwards.